Rename: Hotel.tsx 파일 경로 이동
